DEV : refactoring tests
