+++ src/relife/stochastic_process/tests/test_non_homogeneous_poisson_process.py
@@ -2,13 +2,21 @@
 import numpy as np
 import pytest
 
+from relife.lifetime_model._conditional_model import AgeReplacementModel
+from relife.stochastic_process import NonHomogeneousPoissonProcess
 from relife.stochastic_process._non_homogeneous_poisson_process import NHPPData
+from relife.stochastic_process._sample._iterables import (
+    NonHomogeneousPoissonProcessIterable,
+)
+from relife.utils import get_nb_assets
+
+from .utils import select_from_struct
 
 
 @pytest.fixture
 def nhpp_data_v0():
     return {
-        "ages_at_events": (11, 13, 21, 25, 27),
+        "ages_at_events": np.array([11, 13, 21, 25, 27], dtype=np.float64),
         "events_assets_ids": ("AB2", "CX13", "AB2", "AB2", "CX13"),
     }
 
@@ -16,10 +24,10 @@ def nhpp_data_v0():
 @pytest.fixture
 def nhpp_data_v1():
     return {
-        "ages_at_events": (11, 13, 21, 25, 27),
+        "ages_at_events": np.array([11, 13, 21, 25, 27], dtype=np.float64),
         "events_assets_ids": ("AB2", "CX13", "AB2", "AB2", "CX13"),
-        "first_ages": (10, 12),
-        "last_ages": (35, 60),
+        "first_ages": np.array([10, 12], dtype=np.float64),
+        "last_ages": np.array([35, 60], dtype=np.float64),
         "model_args": (np.array([1.2, 5.5]), np.array([37.2, 22.2])),
         "assets_ids": ("AB2", "CX13"),
     }
@@ -28,7 +36,7 @@ def nhpp_data_v1():
 @pytest.fixture
 def nhpp_data_v2():
     return {
-        "ages_at_events": (11, 13, 21, 25, 27),
+        "ages_at_events": np.array([11, 13, 21, 25, 27], dtype=np.float64),
         "events_assets_ids": ("AB2", "CX13", "AB2", "AB2", "CX13"),
         "model_args": (np.array([1.2, 5.5]), np.array([37.2, 22.2])),
         "assets_ids": ("AB2", "CX13"),
@@ -38,20 +46,20 @@ def nhpp_data_v2():
 @pytest.fixture
 def nhpp_data_v3():
     return {
+        "ages_at_events": np.array([11, 13, 21, 25, 27], dtype=np.float64),
         "events_assets_ids": ("AB2", "CX13", "AB2", "AB2", "CX13"),
         "assets_ids": ("AB2", "CX13"),
-        "first_ages": (10, 12),
+        "first_ages": np.array([10, 12], dtype=np.float64),
         "model_args": (np.array([1.2, 5.5]), np.array([37.2, 22.2])),
-        "ages_at_events": (11, 13, 21, 25, 27),
     }
 
 
 @pytest.fixture
 def nhpp_data_v4():
     return {
-        "ages_at_events": (11, 13, 21, 25, 27),
+        "ages_at_events": np.array([11, 13, 21, 25, 27], dtype=np.float64),
         "events_assets_ids": ("AB2", "CX13", "AB2", "AB2", "CX13"),
-        "last_ages": (35, 60),
+        "last_ages": np.array([35, 60], dtype=np.float64),
         "model_args": (np.array([1.2, 5.5]), np.array([37.2, 22.2])),
         "assets_ids": ("AB2", "CX13"),
     }
@@ -101,95 +109,82 @@ def test_nhhp_data_v4(nhpp_data_v4):
     assert np.all(model_args[1] == np.array([37.2, 37.2, 37.2, 37.2, 22.2, 22.2, 22.2]))
 
 
-#
-# DATA_V0 = {
-#     "assets_ids": np.array(
-#         ["S00", "S00", "S01", "S01", "S02", "S11", "S12", "S12", "S13"], dtype=np.str_
-#     ),
-#     "timeline": np.array(
-#         [
-#             5.01409347,
-#             8.0,
-#             4.0485904,
-#             8.0,
-#             4.0,
-#             8.0,
-#             5.07210737,
-#             8.0,
-#             4.0,
-#         ],
-#         dtype=np.float64,
-#     ),
-#     "ages": np.array(
-#         [5.01409347, 8.0, 4.0485904, 8.0, 4.0, 8.0, 5.07210737, 8.0, 4.0],
-#         dtype=np.float64,
-#     ),
-#     "events_indicators": np.array(
-#         [True, False, True, False, False, False, True, False, False], dtype=np.bool_
-#     ),
-#     "entries": np.array(
-#         [2.0, 5.01409347, 0.0, 4.0485904, 0.0, 2.0, 0.0, 5.07210737, 0.0],
-#         dtype=np.float64,
-#     ),
-# }
-#
-# @pytest.mark.skip(reason="no way of currently testing this")
-# def _test_nhpp_to_failure_data(data):
-#
-#     timeline = data["timeline"]
-#     assets_ids = data["assets_ids"]
-#     ages = data["ages"]
-#     entries = data["entries"]
-#     events_indicators = data["events_indicators"]
-#
-#     sort_ind = np.lexsort((timeline, assets_ids))
-#
-#     entries = entries[sort_ind]
-#     events_indicators = events_indicators[sort_ind]
-#     ages = ages[sort_ind]
-#     assets_ids = assets_ids[sort_ind]
-#
-#     first_ages_index = np.roll(assets_ids, 1) != assets_ids
-#     last_ages_index = np.roll(first_ages_index, -1)
-#
-#     immediatly_replaced = np.logical_and(~events_indicators, first_ages_index)
-#
-#     prefix = np.full_like(assets_ids[immediatly_replaced], "Z", dtype=np.str_)
-#     _assets_ids = np.char.add(prefix, assets_ids[immediatly_replaced])
-#     first_ages = entries[immediatly_replaced].copy()
-#     last_ages = ages[immediatly_replaced].copy()
-#
-#     events_assets_ids = assets_ids[events_indicators]
-#     events_ages = ages[events_indicators]
-#     other_assets_ids = np.unique(events_assets_ids)
-#     _assets_ids = np.concatenate((_assets_ids, other_assets_ids))
-#     first_ages = np.concatenate(
-#         (first_ages, entries[first_ages_index & events_indicators])
-#     )
-#     last_ages = np.concatenate(
-#         (last_ages, ages[last_ages_index & ~immediatly_replaced])
-#     )
-#
-#     return events_assets_ids, events_ages, _assets_ids, first_ages, last_ages
-#
-# @pytest.mark.skip(reason="no way of currently testing this")
-# def test_nhpp_to_failure_data():
-#
-#     events_assets_ids, events_ages, assets_ids, first_ages, last_ages = (
-#         _test_nhpp_to_failure_data(DATA_V0)
-#     )
-#
-#     assert np.all(events_assets_ids == np.array(["S00", "S01", "S12"], dtype=np.str_))
-#     assert np.all(
-#         events_ages == np.array([5.01409347, 4.0485904, 5.07210737], dtype=np.float64)
-#     )
-#     assert np.all(
-#         assets_ids
-#         == np.array(["ZS02", "ZS11", "ZS13", "S00", "S01", "S12"], dtype=np.str_)
-#     )
-#     assert np.all(
-#         first_ages == np.array([0.0, 2.0, 0.0, 2.0, 0.0, 0.0], dtype=np.float64)
-#     )
-#     assert np.all(
-#         last_ages == np.array([4.0, 8.0, 4.0, 8.0, 8.0, 8.0], dtype=np.float64)
-#     )
+def test_basic_sampling(distribution):
+    nhpp = NonHomogeneousPoissonProcess(distribution)
+    nb_samples = 10
+    t0 = 0.0
+    tf = distribution.ppf(0.95)
+
+    iterable = NonHomogeneousPoissonProcessIterable(nhpp, nb_samples, (t0, tf))
+    struct_array = np.concatenate(tuple(iterable))
+    struct_array = np.sort(struct_array, order=("asset_id", "sample_id", "timeline"))
+
+    # Check NHPP property for each sample
+    for i in range(nb_samples):
+        select_sample = select_from_struct(struct_array, sample_id=i)
+        np.testing.assert_equal(select_sample["time"][:-1], select_sample["entry"][1:])
+
+
+def test_age_replacement_sampling(distribution, ar):
+    nhpp = NonHomogeneousPoissonProcess(distribution)
+
+    trial_model = AgeReplacementModel(distribution).freeze(ar)
+    nb_assets = get_nb_assets(*trial_model.args)
+    ar_reshaped = trial_model.args[0]
+
+    nb_samples = 10
+    t0 = distribution.ppf(0.3)
+    tf = 3 * distribution.ppf(0.95)
+
+    iterable = NonHomogeneousPoissonProcessIterable(nhpp, nb_samples, (t0, tf), ar=ar)
+    struct_array = np.concatenate(tuple(iterable))
+    struct_array = np.sort(struct_array, order=("asset_id", "sample_id", "timeline"))
+
+    # Check that all times are less than ar for each asset
+    for i in range(nb_assets):
+        ar_asset = np.atleast_1d(ar_reshaped)[i]
+        select_asset = select_from_struct(struct_array, asset_id=i)
+        assert (select_asset["time"] <= ar_asset + 1e-5).all()
+
+
+def test_regression_sampling(frozen_regression):
+    nhpp = NonHomogeneousPoissonProcess(frozen_regression)
+    nb_assets = frozen_regression.args[0].shape[0]
+    t0 = 0.0
+    tf = frozen_regression.ppf(0.95).min()
+    nb_samples = 10
+
+    iterable = NonHomogeneousPoissonProcessIterable(nhpp, nb_samples, (t0, tf))
+    struct_array = np.concatenate(tuple(iterable))
+    struct_array = np.sort(struct_array, order=("asset_id", "sample_id", "timeline"))
+
+    # Check NHPP proprerty for each sample and each asset
+    for i in range(nb_assets):
+        for j in range(nb_samples):
+            select_sample = select_from_struct(struct_array, asset_id=i, sample_id=j)
+            np.testing.assert_equal(
+                select_sample["time"][:-1], select_sample["entry"][1:]
+            )
+
+
+def test_age_replacement_regression_sampling(frozen_regression, ar):
+    nhpp = NonHomogeneousPoissonProcess(frozen_regression)
+
+    trial_model = AgeReplacementModel(frozen_regression).freeze(ar)
+    nb_assets = get_nb_assets(*trial_model.args)
+    ar_reshaped = trial_model.args[0]
+
+    t0 = frozen_regression.ppf(0.25).min()
+    tf = 10 * frozen_regression.ppf(0.75).max()
+    nb_samples = 10
+
+    iterable = NonHomogeneousPoissonProcessIterable(nhpp, nb_samples, (t0, tf), ar=ar)
+    struct_array = np.concatenate(tuple(iterable))
+    struct_array = np.sort(struct_array, order=("asset_id", "sample_id", "timeline"))
+
+    # check all times are bounded by the age of replacement
+    # add a small constant for numerical approximations
+    for i in range(nb_assets):
+        ar_asset = np.atleast_1d(ar_reshaped)[i]
+        times = select_from_struct(struct_array, asset_id=i)["time"]
+        assert (times <= ar_asset + 1e-5).all()
