+++ src/relife/stochastic_process/tests/utils.py
@@ -1,3 +1,4 @@
+# pyright: basic
 import numpy as np
 from numpy.typing import NDArray
